plug-ins: must faster implementation for tagged disposal mode of animation playback.
In particular, each frame keeps track of which other frames are actually
the same drawable, and even when the finale frame is different, but has
a common base, I "fork" from the base, instead of re-generating it from
the start.
Finally at render time, I don't redraw again the same frame twice on a row.