plug-ins: reorganize caching to have a visible GUI immediately.
Caching can take time and would delay the exposition of the GUI,
having one wonder whether the plugin really started or not. Now the
animation structure (saved in parasite as XML) is immediately parsed,
and the frame caching is only started once the GUI is exposed.
It is still not perfect because caching makes the GUI hardly responsive,
but this is a first step.
Also took the opportunity to some cleaning, renaming (i.e
s/load_xml/deserialize/) and code reorganization here and there.