plug-ins: render the frame buffer in a separate thread.
This way, the GUI does not have to hang forever each time the frames'
composition changes. It still does for changes on many frames at once. I
will have to investigate the issue later.
All GTK+ code still has to stay in the same thread. This is why I emit
the "cache-updated" signal in the main thread as regularly verified by a
separate queue.
Otherwise the plugin would crash with:
> animation-play: Fatal IO error 11 (Resource temporarily unavailable) on X server :0.0.
> [xcb] Unknown request in queue while dequeuing
> [xcb] Most likely this is a multi-threaded client and XInitThreads has not been called
> [xcb] Aborting, sorry about that.
> animation-play: xcb_io.c:165: dequeue_pending_request: Assertion `!xcb_xlib_unknown_req_in_deq' failed.