plug-ins: onion skins can now be configured.
Only up to 5 onion skins are allowed right now (that's already a lot to
not have a messy output). Frame position with identical contents are
overlooked (if we shoot on 2s or 3s, obviously displaying an onion skin
of identical content is meaningless).
This data is not yet saved in the parasite, but this will come.